remove repeated local css
