--- sotopia/ui/streamlit_ui/__init__.py
@@ -1,3 +1,0 @@
-from .utils import reset_database
-
-__all__ = ["reset_database"]
