+++ sotopia/ui/streamlit_ui/pages/display_characters.py
@@ -3,12 +3,10 @@ import streamlit as st
 from sotopia.database import AgentProfile
 
 # Importing avatars
-from sotopia.ui.streamlit_ui.rendering import render_character
+from sotopia.ui.streamlit_ui.rendering import render_character, local_css
 
 
-def local_css(file_name: str) -> None:
-    with open(file_name) as f:
-        st.markdown(f"<style>{f.read()}</style>", unsafe_allow_html=True)
+local_css("./css/style.css")
 
 
 def display_characters() -> None:
